refactor : 컨트롤러에서의 model, view 조작을 service에 위임 및 prettier 일괄 적용
