--- src/listeners.js
@@ -1,6 +1,0 @@
-const addLottoGameListeners = () => {
-/*   $showResultButton.addEventListener('click', onModalShow);
-  $modalClose.addEventListener('click', onModalClose); */
-};
-
-export { addLottoGameListeners };
