+++ src/utils/random.js
@@ -1,3 +1,3 @@
 export const getRandomNumber = (min, max) => {
-    return Math.floor(Math.random() * max + min);
-}
+  return Math.floor(Math.random() * max + min);
+};
